+++ src/connect/flash.js
@@ -3,48 +3,55 @@ const _               = require('lodash');
 const { v4: uuidv4 }  = require('uuid');
 
 const cache           = require('../cache');
+const logger          = require('../logger');
 
 const NS              = 'cacheflash';
+const SIZE_THRESHOLD  = 1024;  // 1KB - auto switch to cacheflash
+const SIZE_WARNING    = 10240; // 10KB - log warning
 
 //
 module.exports = async (req, res, next) => {
 
-  req.session.flash       = req.session.flash || {};
-  res.locals.flash        = req.session.flash;
+  req.session.flash             = req.session.flash || {};
+  res.locals.flash              = req.session.flash;
 
-  req.session.cacheflash  = req.session.cacheflash || [];
-  const cacheflash        = req.session.cacheflash;
+  req.session._igo_cacheflash   = req.session._igo_cacheflash || [];
+  const cacheflash              = req.session._igo_cacheflash;
 
   if (req.method === 'GET') {
     // clear flash scope
-    req.session.flash       = {};
-    req.session.cacheflash  = [];
+    req.session.flash             = {};
+    req.session._igo_cacheflash   = [];
   }
 
   // save flash data in session
-  req.flash = function(key, value) {
+  req.flash = (key, value) => {
+    const size = JSON.stringify(value).length;
+
+    // Auto-switch to cacheflash if object is too large
+    if (size > SIZE_THRESHOLD) {
+      if (size > SIZE_WARNING) {
+        logger.warn(`Flash object "${key}" is very large (${(size/1024).toFixed(1)}KB), automatically using cacheflash. Consider reducing data size.`);
+      }
+      return req.cacheflash(key, value);
+    }
+
     req.session.flash[key] = value;
   };
 
   // save flash data in redis
-  req.cacheflash = function(key, value) {
+  req.cacheflash = (key, value) => {
     const uuid = uuidv4();
-    req.session.cacheflash.push(uuid);
-    const obj = {};
-    obj[key] = value;
-    cache.put(NS, uuid, obj, 60); // 60s
+    req.session._igo_cacheflash.push(uuid);
+    cache.put(NS, uuid, { [key]: value }, 60); // 60s
   };
 
-  if (!cacheflash.length) {
-    return next();
-  }
-
-  if (cacheflash) {
-    // async load cacheflash objects
-    for (const uuid of cacheflash) {
-      const obj = await cache.get(NS, uuid);
-      _.merge(res.locals.flash, obj);
-    }
+  // Load cacheflash objects in parallel
+  if (cacheflash.length) {
+    const objects = await Promise.all(
+      cacheflash.map(uuid => cache.get(NS, uuid))
+    );
+    objects.forEach(obj => _.merge(res.locals.flash, obj));
   }
 
   next();